backend/requirements.txt
aiosqlite>=0.20.0
aiohttp>=3.9.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.0
pydantic-settings>=2.3.0